+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ----------------
 
 * `Feature #17 <https://github.com/SolutionsCloud/apidoc/issues/17>`_ Merge ApiDoc's binaries
+* `Feature #18 <https://github.com/SolutionsCloud/apidoc/issues/18>`_ Merge cli arguments -f and -d in one argument -i (input)
 
 1.1 (2013-08-28)
 ----------------